+++ config/hard/Hangman/code/CGCode.js
@@ -1,0 +1,7 @@
+const word = readline();
+const chars = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');